Add working in background option
Added option to minimize app on closing when setting are set (Minimizin on closing is set in constructor so changing setting will take effect after app restart[tooltip shows it in settings view]). Added new Relay command class with T generic for parameters passed with Closing event. Also in constructor, when setting is set, the app will add NotifyIcon so that app can be closed and opened from system tray.